+++ src/test/java/mmarquee/automation/controls/AutomationWindowTest2.java
@@ -19,18 +19,14 @@ import com.sun.jna.platform.win32.User32;
 import com.sun.jna.platform.win32.Win32Exception;
 import com.sun.jna.platform.win32.WinDef;
 import com.sun.jna.ptr.IntByReference;
-import mmarquee.automation.AutomationElement;
-import mmarquee.automation.AutomationException;
-import mmarquee.automation.ControlType;
-import mmarquee.automation.ElementNotFoundException;
+import mmarquee.automation.*;
 import mmarquee.automation.controls.rebar.AutomationReBar;
 import mmarquee.automation.pattern.ItemContainer;
 import mmarquee.automation.pattern.PatternNotFoundException;
 import mmarquee.automation.pattern.Window;
+import mmarquee.automation.uiautomation.IUIAutomation;
 import mmarquee.automation.uiautomation.IUIAutomationElement3;
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.*;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -49,6 +45,15 @@ import static org.mockito.Mockito.*;
  * AutomationWindow tests that rely on mocking behaviour
  */
 public class AutomationWindowTest2 {
+
+    @BeforeClass
+    public static void checkOs() throws Exception {
+        Assume.assumeTrue(isWindows());
+    }
+
+    private static boolean isWindows() {
+        return System.getProperty("os.name").toLowerCase().contains("windows");
+    }
 
     @Mock
     AutomationElement element;
@@ -69,7 +74,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void testMaximize_Gets_Maximize_From_Pattern() throws AutomationException, PatternNotFoundException {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.maximize();
 
@@ -78,7 +86,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void testMinimize_Calls_Minimize_From_Pattern() throws AutomationException, PatternNotFoundException {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.minimize();
 
@@ -89,7 +100,10 @@ public class AutomationWindowTest2 {
     public void testGetName_Gets_Name_From_Element() throws AutomationException, PatternNotFoundException {
         when(element.getName()).thenReturn("NAME");
 
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         String name = wndw.getName();
 
@@ -98,7 +112,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void test_focus_Calls_setFocus_From_element() throws Exception {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.focus();
 
@@ -107,7 +124,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void test_WaitForIdleInput_Calls_waitForInputIdle_From_Window() throws Exception {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.waitForInputIdle(100);
 
@@ -116,7 +136,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void test_isTopMost_Calls_isTopMost_From_Window() throws Exception {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.isTopMost();
 
@@ -125,7 +148,10 @@ public class AutomationWindowTest2 {
 
     @Test
     public void test_isModal_Calls_isModal_From_Window() throws Exception {
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.isModal();
 
@@ -146,7 +172,10 @@ public class AutomationWindowTest2 {
             }
         }).when(element).getPropertyValue(anyInt());
 
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         wndw.setTransparency(100);
 
@@ -167,7 +196,10 @@ public class AutomationWindowTest2 {
             }
         }).when(element).getPropertyValue(anyInt());
 
-        AutomationWindow wndw = new AutomationWindow(element, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         WinDef.HWND handle = wndw.getNativeWindowHandle();
 
@@ -196,7 +228,10 @@ public class AutomationWindowTest2 {
 
         localElement.element = elem;
 
-        AutomationWindow wndw = new AutomationWindow(localElement, window, container);
+        IUIAutomation mocked_automation = Mockito.mock(IUIAutomation.class);
+        UIAutomation instance = new UIAutomation(mocked_automation);
+
+        AutomationWindow wndw = new AutomationWindow(element, window, container, instance);
 
         WinDef.HWND handle = wndw.getNativeWindowHandle();
