+++ src/test/java/mmarquee/automation/utils/UtilsTest.java
@@ -22,9 +22,7 @@ import com.sun.jna.platform.win32.Tlhelp32;
 import com.sun.jna.platform.win32.User32;
 import com.sun.jna.platform.win32.WinDef;
 import mmarquee.automation.BaseAutomationTest;
-import org.junit.After;
-import org.junit.Ignore;
-import org.junit.Test;
+import org.junit.*;
 import org.junit.runner.RunWith;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
@@ -45,6 +43,14 @@ import static org.mockito.Mockito.verify;
 @PrepareForTest({ Utils.class })
 public class UtilsTest extends BaseAutomationTest {
 
+    @BeforeClass
+    public static void checkOs() throws Exception {
+        Assume.assumeTrue(isWindows());
+    }
+
+    private static boolean isWindows() {
+        return System.getProperty("os.name").toLowerCase().contains("windows");
+    }
     static {
         ClassLoader.getSystemClassLoader().setDefaultAssertionStatus(true);
     }
